+++ src/CameraControls/index.ts
@@ -1,3 +1,4 @@
 export * from './CameraControls';
 export * from './useCameraControls';
 export * from './useCenterGraph';
+export * from './utils';
